# backend/openapi.json
{"openapi": "3.1.0", "info": {"title": "CLANK", "version": "0.1.0"}, "paths": {"/clank/login/access-token": {"post": {"tags": ["login"], "summary": "Login Access Token", "description": "OAuth2 Compatible Token login, get an access token for future requests", "operationId": "login-login_access_token", "requestBody": {"content": {"application/x-www-form-urlencoded": {"schema": {"$ref": "#/components/schemas/Body_login-login_access_token"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Token"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/clank/users/signup/{user_role}": {"post": {"tags": ["users"], "summary": "Create User", "description": "Registers an user into the database as student", "operationId": "users-create_user", "parameters": [{"name": "user_role", "in": "path", "required": true, "schema": {"type": "string", "title": "User Role"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserRegister"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Token"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/clank/users/me": {"get": {"tags": ["users"], "summary": "Get Me User", "description": "Get me info", "operationId": "users-get_me_user", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserPublic"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}, "delete": {"tags": ["users"], "summary": "Delete Me", "description": "Delete me user and say goodbye", "operationId": "users-delete_me", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/clank/users/{user_id}": {"get": {"tags": ["users"], "summary": "Get User", "description": "Get User depending on the given user ID", "operationId": "users-get_user", "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "string", "title": "User Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/clank/users/me/update-password": {"patch": {"tags": ["users"], "summary": "Update Password", "description": "Allows the user to update their password", "operationId": "users-update_password", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserUpdatePassword"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/clank/users/me/update": {"patch": {"tags": ["users"], "summary": "Update Me", "description": "Update me user", "operationId": "users-update_me", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserUpdate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserPublic"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/clank/events": {"post": {"tags": ["events"], "summary": "Create Event", "operationId": "events-create_event", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/EventAdminCreate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/clank/events/pub/list": {"get": {"tags": ["events"], "summary": "List Public Events", "operationId": "events-list_public_events", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/EventList"}, "type": "array", "title": "Response Events-List Public Events"}}}}}}}, "/clank/events/list": {"get": {"tags": ["events"], "summary": "List Events", "operationId": "events-list_events", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/clank/events/random": {"get": {"tags": ["events"], "summary": "Random Button", "operationId": "events-random_button", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/clank/events/get_tickets": {"get": {"tags": ["events"], "summary": "Get Tickets", "operationId": "events-get_tickets", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/clank/events/pub/{event_id}": {"get": {"tags": ["events"], "summary": "Read Public Event", "operationId": "events-read_public_event", "parameters": [{"name": "event_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Event Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/clank/events/{event_id}": {"get": {"tags": ["events"], "summary": "Read Event", "operationId": "events-read_event", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "event_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Event Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["events"], "summary": "Update Event", "operationId": "events-update_event", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "event_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Event Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EventUpdate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "patch": {"tags": ["events"], "summary": "Patch Event", "operationId": "events-patch_event", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "event_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Event Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EventUpdate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["events"], "summary": "Delete Event", "operationId": "events-delete_event", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "event_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Event Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/clank/{event_id}/add_ticket/": {"post": {"tags": ["events"], "summary": "Add Ticket", "operationId": "events-add_ticket", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "event_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Event Id"}}, {"name": "ticket", "in": "query", "required": true, "schema": {"type": "string", "title": "Ticket"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/clank/{event_id}/remove_ticket": {"post": {"tags": ["events"], "summary": "Remove Ticket", "operationId": "events-remove_ticket", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "event_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Event Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/clank/reviews/add": {"post": {"tags": ["events"], "summary": "Add Review", "operationId": "events-add_review", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ReviewAdd"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/clank/reviews/{review_id}": {"delete": {"tags": ["events"], "summary": "Remove Review", "operationId": "events-remove_review", "parameters": [{"name": "review_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Review Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/clank/reviews/event/{event_id}": {"get": {"tags": ["events"], "summary": "Get Event Reviews", "operationId": "events-get_event_reviews", "parameters": [{"name": "event_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Event Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/clank/calendar/": {"get": {"tags": ["calendar"], "summary": "Get User Calendar", "description": "Gets the calendar of the saved events from the current user", "operationId": "calendar-get_user_calendar", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/EventList"}, "type": "array", "title": "Response Calendar-Get User Calendar"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/clank/calendar/{event_id}": {"delete": {"tags": ["calendar"], "summary": "Delete Event Calendar", "operationId": "calendar-delete_event_calendar", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "event_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Event Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["calendar"], "summary": "Save Event Calendar", "operationId": "calendar-save_event_calendar", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "event_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Event Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/clank/tools/users/get-all-users": {"get": {"tags": ["Tools", "Users"], "summary": "Get All Users", "description": "Get all users from the database and returns them (including personal detail)\nScope: \"admin\"", "operationId": "Tools-get_all_users", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/clank/tools/users/get-all-organizers": {"get": {"tags": ["Tools", "Users"], "summary": "Get All Organizers", "description": "Get all organizers from the database and returns them (including personal detail)\nScope: \"admin\"", "operationId": "Tools-get_all_organizers", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/clank/tools/users/delete/{user_id}": {"delete": {"tags": ["Tools", "Users"], "summary": "Delete User", "description": "Directly delete a user from its user_id from the database\nScope: \"admin\"", "operationId": "Tools-delete_user", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "string", "title": "User Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/clank/tools/analytics/user_count": {"get": {"tags": ["Tools", "Analytics"], "summary": "Get Num Users", "description": "Retrieves the number of users on the platform without sending the whole database table\nScope: \"admin\"", "operationId": "Tools-get_num_users", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/clank/tools/analytics/organizer_count": {"get": {"tags": ["Tools", "Analytics"], "summary": "Get Num Organizers", "description": "Retrieves the number of organizers on the platform without sending the whole database table\nScope: \"admin\"", "operationId": "Tools-get_num_organizers", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/clank/tools/analytics/pronoun_count": {"get": {"tags": ["Tools", "Analytics"], "summary": "Get Pronoun Count", "description": "Get the pronouns count from the whole user base\nScope: \"admin\"", "operationId": "Tools-get_pronoun_count", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/clank/tools/analytics/average_age": {"get": {"tags": ["Tools", "Analytics"], "summary": "Get Average Age", "description": "Get the average age of the user base\nScope: \"admin\"", "operationId": "Tools-get_average_age", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/clank/tools/analytics/average_age/{event_id}": {"get": {"tags": ["Tools", "Analytics", "events"], "summary": "Get Event Average Age", "description": "Get the average age of the attendees in a given event\nScope: \"organizer\"", "operationId": "Tools-get_event_average_age", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "event_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Event Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/clank/tools/analytics/attendees-list-export/{event_id}": {"get": {"tags": ["Tools", "Analytics", "events"], "summary": "Get Attendees List", "description": "Export an event list as a csv str to be written unto a file\nScope: \"organizer\"", "operationId": "Tools-get_attendees_list", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "event_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Event Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/clank/tools/analytics/get-all-events/detail": {"get": {"tags": ["Tools", "Analytics", "events"], "summary": "Get All Events", "description": "Get all events' detail\nScope: \"admin\"", "operationId": "Tools-get_all_events", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/clank/tools/moderation/queue": {"get": {"tags": ["Tools"], "summary": "Get Queue", "operationId": "Tools-get_queue", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/ModQueue"}, "type": "array", "title": "Response Tools-Get Queue"}}}}}}, "post": {"tags": ["Tools"], "summary": "Add Report", "operationId": "Tools-add_report", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/AddToQueue"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ModQueue"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/clank/tools/moderation/queue/{req}": {"delete": {"tags": ["Tools"], "summary": "Delete Report", "operationId": "Tools-delete_report", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "req", "in": "path", "required": true, "schema": {"type": "integer", "title": "Req"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ModQueue"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/clank/seed-database": {"post": {"tags": ["Admin Utilities"], "summary": "Seed Database", "operationId": "Admin Utilities-seed_database", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"AddToQueue": {"properties": {"user_id": {"type": "string", "title": "User Id"}, "desc": {"type": "string", "title": "Desc"}}, "type": "object", "required": ["user_id", "desc"], "title": "AddToQueue"}, "Body_login-login_access_token": {"properties": {"grant_type": {"anyOf": [{"type": "string", "pattern": "^password$"}, {"type": "null"}], "title": "Grant Type"}, "username": {"type": "string", "title": "Username"}, "password": {"type": "string", "format": "password", "title": "Password"}, "scope": {"type": "string", "title": "Scope", "default": ""}, "client_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Client Id"}, "client_secret": {"anyOf": [{"type": "string"}, {"type": "null"}], "format": "password", "title": "Client Secret"}}, "type": "object", "required": ["username", "password"], "title": "Body_login-login_access_token"}, "EventAdminCreate": {"properties": {"name": {"type": "string", "title": "Name"}, "description": {"type": "string", "title": "Description"}, "price": {"type": "number", "title": "Price"}, "location": {"type": "string", "title": "Location"}, "start_time": {"type": "string", "format": "date-time", "title": "Start Time"}, "end_time": {"type": "string", "format": "date-time", "title": "End Time"}, "tags": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Tags"}, "pictures": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Pictures"}, "visibility": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Visibility", "default": "private"}, "organizer_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Organizer Id"}, "count_attendees": {"type": "integer", "title": "Count Attendees", "default": 0}, "capacity": {"type": "integer", "title": "Capacity", "default": 1}, "tickets_left": {"type": "integer", "title": "Tickets Left", "default": 1}}, "type": "object", "required": ["name", "description", "price", "location", "start_time", "end_time"], "title": "EventAdminCreate"}, "EventList": {"properties": {"id": {"type": "integer", "title": "Id"}, "name": {"type": "string", "title": "Name"}, "description": {"type": "string", "title": "Description"}, "price": {"type": "number", "title": "Price"}, "location": {"type": "string", "title": "Location"}, "start_time": {"type": "string", "format": "date-time", "title": "Start Time"}, "end_time": {"type": "string", "format": "date-time", "title": "End Time"}, "tags": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Tags"}, "pictures": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Pictures"}, "organizer_id": {"type": "string", "title": "Organizer Id"}}, "type": "object", "required": ["id", "name", "description", "price", "location", "start_time", "end_time", "organizer_id"], "title": "EventList"}, "EventUpdate": {"properties": {"name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Name"}, "description": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Description"}, "price": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Price"}, "location": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Location"}, "start_time": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Start Time"}, "end_time": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "End Time"}, "tags": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Tags"}, "pictures": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Pictures"}, "visibility": {"type": "string", "title": "Visibility"}, "state": {"type": "string", "title": "State"}}, "type": "object", "required": ["visibility", "state"], "title": "EventUpdate"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "Message": {"properties": {"message": {"type": "string", "title": "Message"}}, "type": "object", "required": ["message"], "title": "Message"}, "ModQueue": {"properties": {"req": {"type": "integer", "title": "Req"}, "user_id": {"type": "string", "title": "User Id"}, "desc": {"type": "string", "title": "Desc"}, "date_created": {"type": "string", "format": "date-time", "title": "Date Created"}}, "type": "object", "required": ["req", "user_id"], "title": "ModQueue"}, "ReviewAdd": {"properties": {"user_id": {"type": "string", "title": "User Id"}, "event_id": {"type": "integer", "title": "Event Id"}, "desc": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Desc"}, "star": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Star"}, "date_created": {"type": "string", "format": "date-time", "title": "Date Created"}}, "type": "object", "required": ["user_id", "event_id", "date_created"], "title": "ReviewAdd"}, "Token": {"properties": {"access_token": {"type": "string", "title": "Access Token"}, "token_type": {"type": "string", "title": "Token Type", "default": "bearer"}}, "type": "object", "required": ["access_token"], "title": "Token"}, "UserPublic": {"properties": {"email": {"type": "string", "maxLength": 255, "format": "email", "title": "Email"}, "first_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "First Name"}, "last_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Last Name"}, "pronouns": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Pronouns"}, "username": {"type": "string", "maxLength": 255, "title": "Username"}, "date_of_birth": {"anyOf": [{"type": "string", "format": "date"}, {"type": "null"}], "title": "Date Of Birth"}, "role": {"$ref": "#/components/schemas/UserRole", "default": "student"}, "id": {"type": "string", "title": "Id"}}, "type": "object", "required": ["email", "username", "id"], "title": "UserPublic"}, "UserRegister": {"properties": {"username": {"type": "string", "maxLength": 255, "title": "Username"}, "email": {"type": "string", "maxLength": 255, "format": "email", "title": "Email"}, "password": {"type": "string", "maxLength": 40, "minLength": 8, "title": "Password"}, "first_name": {"type": "string", "maxLength": 255, "title": "First Name"}, "last_name": {"type": "string", "maxLength": 255, "title": "Last Name"}, "pronouns": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Pronouns"}, "date_of_birth": {"anyOf": [{"type": "string", "format": "date"}, {"type": "null"}], "title": "Date Of Birth"}}, "type": "object", "required": ["username", "email", "password"], "title": "UserRegister"}, "UserRole": {"type": "string", "enum": ["admin", "student", "organizer"], "title": "UserRole"}, "UserUpdate": {"properties": {"email": {"anyOf": [{"type": "string", "format": "email"}, {"type": "null"}], "title": "Email"}, "first_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "First Name"}, "last_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Last Name"}, "pronouns": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Pronouns"}}, "type": "object", "title": "UserUpdate"}, "UserUpdatePassword": {"properties": {"current_password": {"type": "string", "maxLength": 40, "minLength": 8, "title": "Current Password"}, "new_password": {"type": "string", "maxLength": 40, "minLength": 8, "title": "New Password"}}, "type": "object", "required": ["current_password", "new_password"], "title": "UserUpdatePassword"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"OAuth2PasswordBearer": {"type": "oauth2", "flows": {"password": {"scopes": {}, "tokenUrl": "/clank/login/access-token"}}}}}}